refactor: Http 클래스 정리
Http 관련 클래스는 model에 있기 보다는 별도의 http 디렉토리에 저장하는게 낫다고 생각하였다.
