+++ src/main/java/http/header/Header.java
@@ -1,0 +1,7 @@
+package http.header;
+
+import java.util.Map;
+
+public abstract class Header {
+    private Map<String, String> properties;
+}
