+++ src/test/java/utils/HttpUtilTest.java
@@ -1,6 +1,6 @@
 package utils;
 
-import model.HttpRequest;
+import http.HttpRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
